Delegate post-request hooks to the repositories instead of cubits
