+++ lib/data/auth/service/local/auth_local_service.dart
@@ -34,10 +34,62 @@ class AuthLocalService {
     }
   }
 
+  Future<String?> get userId {
+    try {
+      return _secureStorage.getUserId();
+    } catch (e, stackTrace) {
+      Flogger.e(
+        "Auth local service error: $e",
+        stackTrace: stackTrace,
+      );
+
+      rethrow;
+    }
+  }
+
+  Future<String?> get userEmail {
+    try {
+      return _secureStorage.getUserEmail();
+    } catch (e, stackTrace) {
+      Flogger.e(
+        "Auth local service error: $e",
+        stackTrace: stackTrace,
+      );
+
+      rethrow;
+    }
+  }
+
   Future<void> saveUserAuthToken(String id) {
     try {
       return _secureStorage.saveUserAuthToken(id);
     } on Exception catch (e, stackTrace) {
+      Flogger.e(
+        "Auth local service error: $e",
+        stackTrace: stackTrace,
+      );
+
+      rethrow;
+    }
+  }
+
+  Future<void> saveUserId(String id) {
+    try {
+      return _secureStorage.saveUserId(id);
+    } catch (e, stackTrace) {
+      Flogger.e(
+        "Auth local service error: $e",
+        stackTrace: stackTrace,
+      );
+
+      rethrow;
+    }
+  }
+
+  Future<void> saveUserEmail(String email) {
+    try {
+      return _secureStorage.saveUserEmail(email);
+    } catch (e, stackTrace) {
       Flogger.e(
         "Auth local service error: $e",
         stackTrace: stackTrace,
