+++ lib/presentation/auth/create_account/bloc/create_account_cubit.dart
@@ -90,8 +90,6 @@ class CreateAccountCubit extends Cubit<CreateAccountState>
         request,
       );
 
-      await _userRepository.saveUserEmail(state.email);
-
       emit(
         CreateAccountStateSuccess(
           createAccountData: createAccountResult,
